+++ yo/__init__.py
@@ -10,6 +10,6 @@ warnings.filterwarnings(
     module="milvus_lite",
 )
 
-__version__ = "0.5.9.0"
+__version__ = "0.5.9.1"
 
 __all__: list[str] = ["__version__"]